system: linux: vfio: Support vfio only for aarch64 architectures
SMMU support is needed for vfio to work. As of now only the
aarch64 architecures like zynqmp/versal has SMMU support.
This patch updates the code to support vfio only for the
architecures with WITH_SMMU flag is enabled. This patch also
fixes the errors that occur when compiling with non-aarch64
architecures like zynq.

By default the WITH_SMMU flag is set to "OFF". This should be
enabled for valid platforms in yocto meta-openamp settings.

Signed-off-by: Anurag Kumar Vulisha <anurag.kumar.vulisha@xilinx.com>
Reviewed-by: Wendy Liang <wendy.liang@xilinx.com>